fix(docs): wrap examples that use useVueFlow in a provider (2.0)
In 2.0's pure-provider model `useVueFlow()` is a pure `inject` consumer — it throws
`USE_VUE_FLOW_OUTSIDE_PROVIDER` when there is no `<VueFlow>` / `<VueFlowProvider>`
ancestor. 13 REPL examples called `useVueFlow()` in their entry `App.vue` setup while
rendering `<VueFlow>` as a child, so the child's `provide` was not visible to the
parent's `inject` and the example threw on load.

Split each into a thin `App.vue` provider wrapper + a `Flow.vue` body:

  <!-- App.vue  --> <VueFlowProvider><Flow /></VueFlowProvider>
  <!-- Flow.vue --> useVueFlow() + <VueFlow>  (reuses the provider's store)

basic, confirm-delete, dnd, helper-lines, intersection, layout, layout-simple,
loopback, stress, transition, update-edge, update-node, validation. Custom
node/edge/control components that call `useVueFlow()` render *inside* `<VueFlow>`,
so they already resolve the store — left as-is.

Also fixes latent 2.0 bugs surfaced once these examples render again:

- layout / layout-simple `useLayout`: `getNode(id).dimensions` → `.measured?` — in
  2.0 `getNode` returns the user node and dimensions live on `measured`.
- layout `AnimationEdge`: `props.data.isAnimating` → `props.data?.isAnimating` —
  edges are stored verbatim, so `data` may be undefined.
- stress: dropped `<MiniMap>` (triggers a core "Maximum recursive updates" at scale,
  tracked separately) and aligned with the maintained `examples/vite` stress
  (`shallowRef` + `v-model`).
- updatable-edge page: `<Repl example="updateEdge">` → `"reconnectEdge"` to match the
  registry key renamed alongside the reconnect* API (the page reference was stale).

Bumps the REPL Vue to 3.5.0. All 13 verified rendering in the docs REPL (Chrome); the
docs build passes.
