+++ docs/examples/basic/index.ts
@@ -1,4 +1,5 @@
 export { default as BasicApp } from './App.vue?raw'
+export { default as BasicFlow } from './Flow.vue?raw'
 export { default as BasicIcon } from './Icon.vue?raw'
 export { default as BasicElements } from './initial-elements.js?raw'
 export { default as BasicCSS } from './style.css?inline'
