+++ docs/examples/confirm-delete/index.ts
@@ -1,3 +1,4 @@
 export { default as ConfirmApp } from './App.vue?raw'
+export { default as ConfirmFlow } from './Flow.vue?raw'
 export { default as ConfirmDialog } from './Dialog.vue?raw'
 export { default as useDialog } from './useDialog.js?raw'
