+++ docs/examples/dnd/App.vue
@@ -1,32 +1,11 @@
 <script setup>
-import { ref } from 'vue'
-import { VueFlow, useVueFlow } from '@vue-flow/core'
-import DropzoneBackground from './DropzoneBackground.vue'
-import Sidebar from './Sidebar.vue'
-import useDragAndDrop from './useDnD'
-
-const { onConnect, addEdges } = useVueFlow()
-
-const { onDragOver, onDrop, onDragLeave, isDragOver } = useDragAndDrop()
-
-const nodes = ref([])
-
-onConnect(addEdges)
+import { VueFlowProvider } from '@vue-flow/core'
+import Flow from './Flow.vue'
 </script>
 
 <template>
-  <div class="dnd-flow" @drop="onDrop">
-    <VueFlow :nodes="nodes" @dragover="onDragOver" @dragleave="onDragLeave">
-      <DropzoneBackground
-        :style="{
-          backgroundColor: isDragOver ? '#e7f3ff' : 'transparent',
-          transition: 'background-color 0.2s ease',
-        }"
-      >
-        <p v-if="isDragOver">Drop here</p>
-      </DropzoneBackground>
-    </VueFlow>
-
-    <Sidebar />
-  </div>
+  <!-- `Flow` calls `useVueFlow()` and renders `<VueFlow>`, so it needs a `<VueFlowProvider>` ancestor to share one store. -->
+  <VueFlowProvider>
+    <Flow />
+  </VueFlowProvider>
 </template>
