+++ docs/examples/helper-lines/index.ts
@@ -1,4 +1,5 @@
 export { default as HelperLinesApp } from './App.vue?raw'
+export { default as HelperLinesFlow } from './Flow.vue?raw'
 export { default as HelperLinesUtils } from './utils.ts?raw'
 export { default as HelperLinesInitialElements } from './initialElements.ts?raw'
 export { default as HelperLinesStyle } from './style.css?inline'
