+++ docs/examples/intersection/index.ts
@@ -1,2 +1,3 @@
 export { default as IntersectionApp } from './App.vue?raw'
+export { default as IntersectionFlow } from './Flow.vue?raw'
 export { default as IntersectionCSS } from './style.css?inline'
