+++ docs/examples/layout-simple/index.ts
@@ -1,4 +1,5 @@
 export { default as SimpleLayoutApp } from './App.vue?raw'
+export { default as SimpleLayoutFlow } from './Flow.vue?raw'
 export { default as SimpleLayoutElements } from './initial-elements.js?raw'
 export { default as useSimpleLayout } from './useLayout.js?raw'
 export { default as SimpleLayoutIcon } from './Icon.vue?raw'
