+++ docs/examples/layout/index.ts
@@ -1,4 +1,5 @@
 export { default as LayoutApp } from './App.vue?raw'
+export { default as LayoutFlow } from './Flow.vue?raw'
 export { default as LayoutElements } from './initial-elements.js?raw'
 export { default as LayoutNode } from './ProcessNode.vue?raw'
 export { default as LayoutEdge } from './AnimationEdge.vue?raw'
