+++ docs/examples/layout/useLayout.js
@@ -33,7 +33,7 @@ export function useLayout() {
         continue
       }
 
-      dagreGraph.setNode(node.id, { width: graphNode.dimensions.width || 150, height: graphNode.dimensions.height || 50 })
+      dagreGraph.setNode(node.id, { width: graphNode.measured?.width || 150, height: graphNode.measured?.height || 50 })
     }
 
     for (const edge of edges) {
