+++ docs/examples/loopback/index.ts
@@ -1,3 +1,4 @@
 export { default as LoopbackApp } from './App.vue?raw'
+export { default as LoopbackFlow } from './Flow.vue?raw'
 export { default as LoopbackEdge } from './LoopbackEdge.vue?raw'
 export { default as LoopbackCSS } from './style.css?inline'
