+++ docs/examples/stress/index.ts
@@ -1,3 +1,4 @@
 export { default as StressApp } from './App.vue?raw'
+export { default as StressFlow } from './Flow.vue?raw'
 export { default as StressUtils } from './utils.js?raw'
 export { default as StressCSS } from './style.css?inline'
