+++ docs/examples/transition/index.ts
@@ -1,3 +1,4 @@
 export { default as TransitionApp } from './App.vue?raw'
+export { default as TransitionFlow } from './Flow.vue?raw'
 export { default as TransitionEdge } from './TransitionEdge.vue?raw'
 export { default as TransitionCSS } from './style.css?inline'
