+++ docs/examples/update-edge/index.ts
@@ -1,1 +1,2 @@
 export { default as UpdateEdgeApp } from './App.vue?raw'
+export { default as UpdateEdgeFlow } from './Flow.vue?raw'
