+++ docs/examples/validation/index.ts
@@ -1,4 +1,5 @@
 export { default as ValidationApp } from './App.vue?raw'
+export { default as ValidationFlow } from './Flow.vue?raw'
 export { default as ValidationCustomInput } from './CustomInput.vue?raw'
 export { default as ValidationCustomNode } from './CustomNode.vue?raw'
 export { default as ValidationCSS } from './style.css?inline'
